requirements_backend.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.4

# ETL and Data Pipeline
apache-airflow==3.0.1

# Build tools
setuptools>=69.0.0
wheel>=0.42.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Search Engine
elasticsearch==8.17.0

# Analytics Database
clickhouse-driver==0.2.9
clickhouse-sqlalchemy==0.3.2
numpy>=1.26.0

# Cache and Message Queue
redis==5.2.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP Clients
httpx==0.28.1
requests==2.32.3
aiohttp==3.10.11

# Logging
loguru==0.7.3

# Configuration
pydantic-settings==2.7.0
python-dotenv==1.0.1

# File Handling
aiofiles==24.1.0

# Audio Processing
python-magic==0.4.27
mutagen==1.47.0

# Additional Dependencies
starlette==0.41.3
typing-extensions==4.12.2
python-dateutil==2.9.0
orjson==3.10.12

# S3 and Cloud Storage
boto3==1.38.29
botocore==1.38.29

# Data Processing
pandas==2.2.3

# Production server
gunicorn==23.0.0
